add custom loader
